Split events off into their own file.
